+++ packages/router/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* Omit undefined optional params from match params
+
 ## 2.0.1
 
 * The `div` inserted by `announce` sets `top: 0;` to prevent overflow on fixed size websites.